Boost code coverage, reorganize tests
